Make all t-test plots driven by the x coordinate, add TTest$plotRanges
Since so many plots needed to be computed by X instead of Y, I decided
to make all t-test plot generation the same. In addition, to facilitate
projects with multiple lines per plot, I created the plotRanges function
which is used by TTestPlotDataAction to create a maximum plot range.
This allows all the lines to properly run to the edge of the plot area.